feat(#5) : Atomic 파일 구조로 변경
